+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 2. Updated to _modern Go_ with `go fix`.
 3. _docker-dev_ container now uses HiveMQ public MQTT broker.
 4. Migrated from _github.com/uhppoted/uhppoted-mqtt_ to _codeberg.org/uhppoted/uhppoted-mqtt_.
+5. Added `mqtt.PIN.enabled` and `mqtt.firstcard.enabled` to _uhppoted.conf_.
+6. Added (optional) first-card privileges to `get-card` and `put-card` APIs.
 
 
 ## [0.9.0](https://github.com/uhppoted/uhppoted-mqtt/releases/tag/v0.9.0) - 2026-01-27